requirements.txt
# ===========================================
# WildGuard AI - Python Dependencies
# ===========================================
# Install all: pip install -r requirements.txt
# Python version: 3.10 - 3.12 recommended
# ===========================================

# --- Core Data Processing ---
pandas>=2.0.0
numpy>=1.24.0

# --- Machine Learning ---
scikit-learn>=1.3.0
xgboost>=2.0.0
statsmodels>=0.14.0

# --- Dashboard & Visualization ---
streamlit>=1.28.0
plotly>=5.18.0
pydeck>=0.8.0

# --- External APIs & Services ---
groq>=0.9.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
pymongo>=4.6.0
dnspython>=2.6.0

# --- Training Scripts Only (not needed to run dashboard) ---
matplotlib>=3.7.0
seaborn>=0.13.0

# --- Optional: Only needed if running model_comparison.py or prophet_forecasting.py ---
# prophet>=1.1.5

# --- Optional: Only needed if running lstm_trend_detection.py or model_comparison.py ---
# tensorflow>=2.15.0
